feat(domain): add shared ProblemDetail type and wire API client
- Add ProblemDetail interface (RFC 9457) to @genshin/domain
- Use satisfies ProblemDetail in API error handlers (apps/api/src/app.ts)
- Return unknown from API client instead of unsafe generic casts
- Callers narrow at domain boundary with assertion helpers
